Function added to ExpenseDAOimpl

+++ src/main/java/org/example/demo1/domain/models/Expenses.java
@@ -1,5 +1,6 @@
 package org.example.domain.models;
 
+import java.math.BigDecimal;
 import java.time.LocalDate;
 import java.time.LocalDateTime;
 
@@ -33,7 +34,7 @@ public class Expenses {
         this.user_id = user_id;
     }
 
-    public double getAmount() {
+    public BigDecimal getAmount() {
         return amount;
     }
